pyproject.toml
[project]
name = "tdd-with-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "pytest-django>=4.11.1",
    "pytest-playwright>=0.7.2",
]
